refactor : 컴포넌트 각 파일들로 분리 및 컴포넌트 명 변경
