+++ src/app/_component/common/Input/InputInnerBox.tsx
@@ -1,0 +1,12 @@
+interface InputInnerBox {
+  className?: string;
+}
+
+const InputInnerBox = ({
+  className,
+  children,
+}: React.PropsWithChildren<InputInnerBox>) => {
+  return <div className={`${className} flex`}>{children}</div>;
+};
+
+export default InputInnerBox;
