+++ src/app/_component/common/Input/Label.tsx
@@ -1,0 +1,12 @@
+interface LabelProps {
+  className?: string;
+}
+
+const Label = ({
+  className,
+  children,
+}: React.PropsWithChildren<LabelProps>) => {
+  return <label className={className}>{children}</label>;
+};
+
+export default Label;
