+++ src/app/_component/common/Input/SubmitButton.tsx
@@ -1,0 +1,12 @@
+interface SubmitButtonProps {
+  className?: string;
+}
+
+const SubmitButton = ({
+  className,
+  children,
+}: React.PropsWithChildren<SubmitButtonProps>) => {
+  return <label className={className}>{children}</label>;
+};
+
+export default SubmitButton;
