+++ src/app/_component/common/Input/index.tsx
@@ -1,91 +1,25 @@
-import { cn } from "@/utils/cn";
-import React, { ReactNode } from "react";
-import { VariantProps } from "class-variance-authority";
-import { InputVariants } from "./Input.variants";
+import React from "react";
 
-interface InputProps extends VariantProps<typeof InputVariants> {
-  block?: boolean;
-  invalid?: boolean;
-  required?: boolean;
-  disabled?: boolean;
-  wrapperStyle?: React.CSSProperties;
-  placeHolder?: string;
-  type?: string;
-  className: string;
-}
+import Label from "./Label";
+import InputInnerBox from "./InputInnerBox";
+import SubmitButton from "./SubmitButton";
+import InputForm from "./InputForm";
 
 interface InputWrapperProps {
   className?: string;
 }
 
-interface OuterLabelProps {
-  className?: string;
-}
-
-interface SubmitButtonProps {
-  className?: string;
-}
-
-const InputArea = ({
-  block = false,
-  invalid = false,
-  required = false,
-  disabled = false,
-  type,
-  verticalAlign = "center",
-  fontSize,
-  placeHolder = "가나다라마바사",
-  wrapperStyle,
-  className,
-  ...props
-}: InputProps) => {
-  return (
-    <div style={wrapperStyle}>
-      <input
-        className={cn(InputVariants({ verticalAlign }), className)}
-        type={type}
-        placeholder={placeHolder}
-        required={required}
-        disabled={disabled}
-        {...props}
-      />
-    </div>
-  );
-};
-
 const InputWrapper = ({
   className,
-  children,
 }: React.PropsWithChildren<InputWrapperProps>) => {
   return <div className={`${className} flex`}>{children}</div>;
-};
-
-const InputBox = ({
-  className,
-  children,
-}: React.PropsWithChildren<InputWrapperProps>) => {
-  return <div className={`${className} flex`}>{children}</div>;
-};
-
-const Label = ({
-  className,
-  children,
-}: React.PropsWithChildren<OuterLabelProps>) => {
-  return <label className={className}>{children}</label>;
-};
-
-const SubmitButton = ({
-  className,
-  children,
-}: React.PropsWithChildren<SubmitButtonProps>) => {
-  return <label className={className}>{children}</label>;
 };
 
 const Input = Object.assign(InputWrapper, {
-  Label: Label,
-  InputArea: InputArea,
-  InputBox: InputBox,
-  SubmitButton: SubmitButton,
+  Label,
+  InputInnerBox,
+  SubmitButton,
+  InputForm,
 });
 
 export default Input;
