+++ src/app/page.tsx
@@ -7,7 +7,6 @@ export default function Home() {
       <Input>
         <Input.Label>라벨</Input.Label>
         <Input.InputBox>
-          <Input.Label>라벨</Input.Label>
           <Input.InputArea className="w-24" />
         </Input.InputBox>
       </Input>
